Fix IndexError on Jython (#250)
Jython has an unfortunate issue where the socketaddr object isn't
really a tuple, but is instead an item that acts like a tuple under
some circumstances. This commit makes httplib2 convert it to an actual
tuple, so that it will work as expected.

This solution isn't very nice, but without upstream fixes to Jython, this
is the only obvious solution.